backend/requirements.txt
flask==2.2.2
werkzeug==2.2.3
flask-cors==3.0.10
text2emotion==0.0.5
spotipy==2.22.1
requests==2.31.0
python-dotenv==1.0.0
nltk==3.8.1